+++ src/main/kotlin/bot/tx/wsure/top/component/bililiver/VtrGiftLotteryNotify.kt
@@ -28,13 +28,13 @@ class VtrGiftLotteryNotify(val config:List<ChannelConfig>, val client: UnOfficia
             it.value
         }
         config.onEach {
-            if(rale || guilds.contains(it.guildId)){
+            if(rale || guilds.contains(it.guildId.toLong())){
                 client.sender.sendGuildChannelMsgAsync(unofficialGuildMessage(msg,it))
             }
         }
     }
 
     fun unofficialGuildMessage(msg:String,channel: ChannelConfig): SendGuildChannelMsg {
-        return SendGuildChannelMsg( channel.guildId,channel.channelId,msg )  //6000051636714649,1370732,msg))
+        return SendGuildChannelMsg( channel.guildId.toLong(),channel.channelId.toLong(),msg )  //6000051636714649,1370732,msg))
     }
 }
